+++ Patches/ItemDropshipPatch.cs
@@ -1,0 +1,18 @@
+﻿using GeneralImprovements.Utilities;
+using HarmonyLib;
+
+namespace GeneralImprovements.Patches
+{
+    internal static class ItemDropshipPatch
+    {
+        [HarmonyPatch(typeof(ItemDropship), nameof(Start))]
+        [HarmonyPostfix]
+        private static void Start(ItemDropship __instance)
+        {
+            if (Plugin.ShowDropshipOnScanner.Value)
+            {
+                ItemHelper.CreateScanNodeOnObject(__instance.gameObject, 0, 5, 50, "Dropship", 4);
+            }
+        }
+    }
+}