+++ Utilities/NetworkHelper.cs
@@ -13,15 +13,17 @@ namespace GeneralImprovements.Utilities
         }
 
         [ClientRpc]
-        public void SyncExtraDataOnConnectClientRpc(int totalDays, int totalDeaths, int daysSinceLastDeath, ClientRpcParams clientParams)
+        public void SyncExtraDataOnConnectClientRpc(int quotaNum, int totalDays, int totalDeaths, int daysSinceLastDeath, ClientRpcParams clientParams)
         {
             if (!IsServer && StartOfRound.Instance != null)
             {
                 Plugin.MLS.LogInfo("Client received extra data sync RPC.");
+                TimeOfDay.Instance.timesFulfilledQuota = quotaNum;
                 StartOfRound.Instance.gameStats.daysSpent = totalDays;
                 StartOfRound.Instance.gameStats.deaths = totalDeaths;
                 StartOfRoundPatch.DaysSinceLastDeath = daysSinceLastDeath;
 
+                MonitorsHelper.UpdateTotalQuotasMonitors();
                 MonitorsHelper.UpdateTotalDaysMonitors();
                 MonitorsHelper.UpdateDeathMonitors();
             }